Laboratorio/OmniLatam/RETO/requirements.txt
Django==2.0.2
django-admin-rangefilter==0.8.1
django-jet==1.0.8
djangorestframework==3.11.2
psycopg2-binary==2.8.6
pytz==2021.1
